TL-16386 dashboard: Fixed dashboard reset with deleted users
Reviewed-on: https://review.totaralms.com/17575
Tested-by: Jenkins Automation <jenkins@totaralms.com>

+++ totara/dashboard/lib.php
@@ -592,20 +592,23 @@ class totara_dashboard {
 
         $pageid = $this->get_user_pageid($userid);
         if ($pageid) {
-            $context = context_user::instance($userid);
-            if ($blocks = $DB->get_records('block_instances', array('parentcontextid' => $context->id,
-                    'pagetypepattern' => 'my-totara-dashboard-' . $this->id))) {
-                foreach ($blocks as $block) {
-                    if (is_null($block->subpagepattern) || $block->subpagepattern == $pageid) {
-                        blocks_delete_instance($block);
+            $context = context_user::instance($userid, IGNORE_MISSING);
+            if ($context) {
+                if ($blocks = $DB->get_records('block_instances', array('parentcontextid' => $context->id,
+                        'pagetypepattern' => 'my-totara-dashboard-' . $this->id))) {
+                    foreach ($blocks as $block) {
+                        if (is_null($block->subpagepattern) || $block->subpagepattern == $pageid) {
+                            blocks_delete_instance($block);
+                        }
                     }
                 }
+                $DB->delete_records('block_positions', array(
+                    'contextid' => $context->id,
+                    'pagetype' => 'my-totara-dashboard-' . $this->id,
+                    'subpage' => $pageid
+                ));
             }
-            $DB->delete_records('block_positions', array(
-                'contextid' => $context->id,
-                'pagetype' => 'my-totara-dashboard-' . $this->id,
-                'subpage' => $pageid
-            ));
+
             $DB->delete_records('totara_dashboard_user', array('id' => $pageid));
         }
     }
